+++ src/aks-preview/HISTORY.md
@@ -2,9 +2,15 @@
 
 Release History
 ===============
+0.4.73
++++++
+* Update to use 2020-12-01 api-version
+* Add argument '--enable-encryption-at-host'
+
 0.4.72
 ++++++
 * Add --no-uptime-sla
+* Create MSI clusters by default.
 
 0.4.71
 ++++++